Moved Stats Hit code to separate class file Implemented composer based autoloading
