--- includes/class-ssp-stats-hit.php
@@ -1,5 +1,0 @@
-<?php
-
-class SSP_Stats_Hit {
-
-}
